Revert "Attention tracker"
