+++ .gitignore
@@ -5,8 +5,6 @@ reports/
 config.py
 .vscode/
 *.csv
-test-data/
-
 # Dependencies
 node_modules/
 /.pnp
