--- computer-vision_integration/requirements.txt
@@ -1,8 +1,0 @@
-opencv-python>=4.8.0
-mediapipe>=0.10.8
-numpy>=1.23.5
-pandas>=2.0.3
-easyocr>=1.7.1
-torch>=2.0.0
-torchvision>=0.15.0
-Pillow>=10.0.0
